refactor: auto-split test_spatial_algebra_comprehensive.py
